Transport S8 (media store): the /ws media twins commit through the shared store fns
The media/frame-registry store core moves to its neutral home,
web_gateway::media_store (transport-unification S8): HQ-frame
registration (register_dashboard_media_frame), presence-video
register+record (register_presence_video_frame — the ControlRuntime
read stays tunnel-side as a thin edge wrapper), the annotation/clip
context injections, and the DashboardMediaClipOperation type (fields
widened pub(crate) for the /ws accumulator; dashboard_control re-exports
keep every existing reference compiling).

The /ws twins (video_frame, annotation_attach, annotation_submit,
clip_start/frame/end) stop duplicating the store logic inline and call
the same fns the tunnel's api_media_* residue handlers call. Wire
frames, response shapes, and the /ws lane's PresenceLog wordings are
byte-identical; the only observable delta is stderr wording on the
video_frame lane (the shared fn's more descriptive 'presence video …'
prefixes replace the ws arm's bare 'frame registry write failed' /
'camera recording start failed' lines — log-tee only, not wire).

Store behavior notes pinned by the shared fns: register-then-record
order, record only when enabled, auto-start once per stream when ffmpeg
is present (RecordingStarted broadcast on success), feed even after a
failed start, clip accumulation at the data level (register only after
a successful decode) — all exactly as both lanes already behaved.
